fix: moved schemas generation to a separate command and file
